Fix ubsan error on unittest
The ReadFailureFails unittest is passing a nullptr as the buffer but
since the requested size to do IO operation is large, at some point it
hits the code where there is a pointer conversion + addition (on a
nullptr) and that is causing this error. So to fix this, basically,
pass a semi-valid pointer as the buffer instead of the nullptr. This
unit test is not testing the buffer itself anyway.

Bug: crbug.com/1047842
Test: USE="ubsan" FEATURES=test emerge-amd64-generic bsdiff

Change-Id: I7e4d4633953385a960e1ce87fd0c0bfa886b3df7