Include algorithm for suffix_array_index.cc
This is required due to the use of std::mismatch. This
file may have compiled before due to a transitive inclusion
of the <algorithm> header. However, recent changes to LLVM ToT
will prevent these transitive inclusions from happening.

We identified this problem due to emerging bsdiff with `llvm-next`
in ChromeOS.

Bug: 228085014
Test: setup_board --board=amd64-generic \
  && repo download -b change-3564896 \
     chromiumos/overlays/chromiumos-overlay 3564896 \
  && sudo USE='llvm-next -thinlto' emerge \
     sys-devel/llvm sys-libs/libcxx \
  && sudo emerge bsdiff
Change-Id: Ie3af357377c9cf4b51f0d7e1d9758a74605c7b46